+++ BankApp/BankAppTests/Login/LoginRouterTests.swift
@@ -1,0 +1,34 @@
+//
+//  LoginRouterTest.swift
+//  BankAppTests
+//
+//  Created by Victor Hugo Martins Lisboa on 02/10/19.
+//  Copyright © 2019 Victor Lisboa. All rights reserved.
+//
+
+import XCTest
+@testable import BankApp
+
+class LoginRouterTests: XCTestCase {
+
+    override func setUp() {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testRouter() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
